pyproject.toml
[project]
name = "cosplay-photography-toolkit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"

[dependency-groups]
dev = [
    "boto3-stubs~=1.42.59",
    "djangorestframework-stubs>=3.16.8",
]
web = [
    "django>=6.0.2",
    "django-environ>=0.13.0",
    "psycopg2-binary>=2.9.11",
    "djangorestframework>=3.15",
    "boto3>=1.35",
    "nh3>=0.3.3",
]
cli = [
    "click>=8.1",
    "boto3>=1.35",
    "httpx>=0.28",
]

[project.scripts]
photo-tool = "cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli"]
